requirements.txt
torch==2.0.1
transformers==4.30.2
sentence-transformers==2.2.2
huggingface_hub==0.14.1

pandas
numpy
PyMuPDF
scikit-learn
streamlit
plotly
llama-cpp-python
faiss-cpu